requirements.txt
requests==2.7.0
discord.py==1.3.3
